added the name id and name in a separate component, added more comments to code for clarty
