+++ src/components/pokeNameID.css
@@ -1,0 +1,3 @@
+.container {
+    margin: 2px;
+}
