+++ src/containers/pokemonProfile.js
@@ -1,27 +1,15 @@
 import React from 'react';
 import { NavBar } from '../components/navBar';
 import { MovesTabs } from '../components/movesTabs';
+import { PokeNameID } from '../components/pokeNameID'
 import { BaseStats } from '../components/baseStats';
-import { Row, Col, Container } from 'reactstrap';
 import './pokemonProfile.css'
 
 const PokemonProfile = props => {
     const profileStats = props.profile.stats;//an array of stats objects
-    const name = props.data[0]//lowered cased pokemon name
-    const upperCase = name[0].toUpperCase();
-    const lowerCase = name.slice(1);
-    const pokeName = upperCase.concat('', lowerCase);
-    const pokeTitle = props.data[1] + '  -  ' + pokeName;
     return <React.Fragment>
         <NavBar data={props.data} onClick={props.click} />
-        <div>
-            <Row className="float-right">
-                <Col className="col-auto - variable width content" >
-                    <h2 className="title" >{pokeTitle}</h2>
-                </Col>
-            </Row>
-        </div>
-
+        <PokeNameID data={props.data}></PokeNameID>     
         <BaseStats profileData={profileStats}></BaseStats>
         <MovesTabs movesData={props.movesData} onClick={props.movesclick} />
     </React.Fragment>
